♻️  task_shaping.py 추출 — task_type 분류 + write-intent (P0-L step 4)
변경 이유
- `_suggest_task_type` 는 P0-J (#145) 에서 stack_detector 우선 / 키워드
  fallback 으로 정리됐고, 향후 task type 추가 / 키워드 보정이 잦은 영역.
  단일 모듈로 분리해 변경 표면을 명확히 한다.

주요 변경 사항
- 새 파일 `engineering_conversation/task_shaping.py`:
  - `_looks_like_multiple_tasks`
  - `_looks_like_write_request`
  - `_TASK_TYPE_KEYWORDS` 8 종 키워드 테이블
  - `_suggest_task_type` (stack_detector 우선 / 키워드 fallback)
- `_legacy.py` 의 해당 본문 블록 (약 120 줄) 제거 → re-export 만 남김.

비고
- 4657 PASS + 1 skipped 회귀 유지 (동작 변경 0).
- `_normalize` / `split_task_branches` 는 intent_detection 책임이라
  task_shaping 안에서 `from ._legacy import` 로 지연 import (다음 step 5
  에서 intent_detection 추출 시 정리).
- 다음 step (5): status_responses.py 추출.
